requirements.txt
python-dotenv>=1.1.1
groq>=0.31.1